+++ shared/manage-issue-labels.js
@@ -3,12 +3,16 @@ const { logger } = require('./format-log-messages');
 
 /**
  * Adds labels to a specified issue
- * @param {Number} issueNum   -an issue's number
- * @param {Array} labels      -an array containing the labels to add (captures the rest of the parameters)
+ * @param {Object} github      -the octokit instance
+ * @param {Object} context     -the GitHub Actions context object
+ * @param {Object} config      -configuration object
+ * @param {Number} issueNum    -an issue's number
+ * @param {Object} issueLog    -logger object for the specific issue
+ * @param {Array} labels       -an array containing the labels to add (captures the rest of the parameters)
  */
-async function addLabels(github, context, config, issueNum, ...labelsToAdd) {
+async function addLabels(github, context, config, issueNum, issueLog, ...labelsToAdd) {
   if (config.dryRun) {
-    logger.debug(` Would add '${labelsToAdd}' to issue #${issueNum}`);
+    issueLog.info(`  DEBUG: Would add '${labelsToAdd}'`);
     return;
   }
   try {
@@ -19,10 +23,10 @@ async function addLabels(github, context, config, issueNum, ...labelsToAdd) {
       issue_number: issueNum,
       labels: labelsToAdd,
     });
-    logger.info(`  '${labelsToAdd}' label has been added`);
+    issueLog.info(`  '${labelsToAdd}' label has been added`);
     // If an error is found, the rest of the script does not stop.
   } catch (err) {
-    logger.error(`Issue #${issueNum}: Failed to add '${labelsToAdd}': ${err}`);
+    issueLog.info(`  WARNING: failed to add '${labelsToAdd}': ${err}`);
   }
 }
 
@@ -30,13 +34,17 @@ async function addLabels(github, context, config, issueNum, ...labelsToAdd) {
 
 /**
  * Removes labels from a specified issue
- * @param {Number} issueNum    - an issue's number
+ * @param {Object} github      -the octokit instance
+ * @param {Object} context     -the GitHub Actions context object
+ * @param {Object} config      -configuration object
+ * @param {Number} issueNum    -an issue's number
+ * @param {Object} issueLog    -logger object for the specific issue
  * @param {Array} labels       - an array containing the labels to remove (captures the rest of the parameters)
  */
-async function removeLabels(github, context, config, issueNum, ...labelsToRemove) {
+async function removeLabels(github, context, config, issueNum, issueLog, ...labelsToRemove) {
   for (let label of labelsToRemove) {
     if (config.dryRun) {
-      logger.debug(`  Would remove '${label}' from issue #${issueNum}`);
+      issueLog.info(`  DEBUG: Would remove '${label}'`);
       continue;
     }
     try {
@@ -47,10 +55,10 @@ async function removeLabels(github, context, config, issueNum, ...labelsToRemove
         issue_number: issueNum,
         name: label,
       });
-      logger.info(`  '${label}' label has been removed`);
+      issueLog.info(`  '${label}' label has been removed`);
     } catch (err) {
       if (err.status !== 404) {
-        logger.error(`Issue #${issueNum} failed to remove '${label}': ${err}`);
+        issueLog.info(`  WARN: failed to remove '${label}': ${err}`);
       }
     }
   }